Code update for Phase 2 (T5)
Added a check for "item[name]" being null (there was at least 1 entry causing an error because it was a blank item with no name and no stats (but had other attributes)

Added the new Tier Token pieces to the `tier_dict()` dictionary. Now includes all the "{ } of Vanquished { }" which are the new phase 2 tokens. (Phase 1 tokens were "{ } of Fallen { }")